requirements.txt
# CS-GY 6083 - Project Part II
# Streaming Platform Web Application Dependencies

Flask==3.0.0
mysql-connector-python==8.2.0
bcrypt==4.1.2
python-dotenv==1.0.0
Werkzeug==3.0.1
